Add carrent state of Lab2

+++ OAiP_Lab2/task3.2.cpp
@@ -9,7 +9,7 @@ using namespace std;
 
 int main()
 {
-    float X,Y;
+    double X,Y;
     cin >> X >> Y;
     bool a = X > Y;
     a ? cout << X : cout << Y;
